Update(Fibonacci): Change the type of data for the variables in the method fib
Se cambio el tipo de datos de las variables i y j con el del metodo a unsigned log para poder almacenar una mayor cantidad de datos. Tambien se agregaron unas librerias a utilizar.

+++ Fibonacci.c
@@ -1,10 +1,14 @@
 #include <stdio.h>
 #include <stdlib.h>
 #include <omp.h>
+#include <math.h>
+#include <stdbool.h>
 
-int fib(int n){
-    int i;
-    int j;
+// Metodo para realizar el calculo Fibunacci
+unsigned long long fib(unsigned long long n){
+    // Cambiamos variables int por unsigned long para almacenar enteros largos
+    unsigned long i;
+    unsigned long j;
     if(n<2)
     {
         return n;
